Updated Laundrymachine
Added code to handle Input
Added code to handle multiple Centipede issue
Changed Debug messages to be more readable

+++ Laundromat/MyTimer.h
@@ -12,8 +12,8 @@ class MyTimer
   void NewTimer(int Time);
 
   private:
-  int StartingTime;
-  int EndingTime;
+  unsigned long StartingTime;
+  unsigned long EndingTime;
 };
 
 #endif
